added the receipt extension code to a service and removed unused variables
